feat: stabilize cli, handle errors more gracefully (wip)

+++ src/assets.rs
@@ -1,7 +1,20 @@
 use path::installed_path;
 use serde::{Deserialize, Serialize};
+use thiserror::Error;
 
 use crate::traits::ReadSeek;
+
+#[derive(Error, Debug)]
+pub enum AssetError {
+    #[error("Invalid asset structure. No addons folder was identified for asset with id {0}")]
+    InvalidAssetStructure(String),
+    #[error(transparent)]
+    Io(#[from] std::io::Error),
+    #[error(transparent)]
+    Zip(#[from] zip::result::ZipError),
+    #[error("Asset {0} is not installed")]
+    NotInstalled(String),
+}
 
 #[derive(Deserialize, Serialize, Clone, Debug, PartialEq)]
 pub struct AssetInfo {
@@ -31,9 +44,12 @@ pub struct AssetArchive {
 }
 
 impl AssetArchive {
-    pub fn get_plugin_name_and_files_to_extract(&self) -> (String, Vec<String>) {
-        let (plugin_name, plugin_path) =
-            self.get_plugin_info().expect("can find plugin folder path");
+    pub fn get_plugin_name_and_files_to_extract(
+        &self,
+    ) -> Result<(String, Vec<String>), AssetError> {
+        let Some((plugin_name, plugin_path)) = self.get_plugin_info() else {
+            return Err(AssetError::InvalidAssetStructure(self.id.to_string()));
+        };
 
         let file_paths = self
             .archive
@@ -42,7 +58,7 @@ impl AssetArchive {
             .map(String::from)
             .collect();
 
-        (plugin_name, file_paths)
+        Ok((plugin_name, file_paths))
     }
 
     pub fn get_plugin_info(&self) -> Option<(String, String)> {
@@ -83,8 +99,9 @@ impl AssetArchive {
     }
 }
 
-pub fn install(asset_archive: AssetArchive) -> Result<String, std::io::Error> {
-    let (plugin_name, zip_paths_to_extract) = asset_archive.get_plugin_name_and_files_to_extract();
+pub fn install(asset_archive: AssetArchive) -> Result<String, AssetError> {
+    let (plugin_name, zip_paths_to_extract) =
+        asset_archive.get_plugin_name_and_files_to_extract()?;
 
     let mut archive = asset_archive.archive;
 
@@ -110,11 +127,14 @@ pub fn install(asset_archive: AssetArchive) -> Result<String, std::io::Error> {
     Ok(plugin_name)
 }
 
-pub fn uninstall(asset: &AssetInfo) -> Result<(), std::io::Error> {
+pub fn uninstall(asset: &AssetInfo) -> Result<(), AssetError> {
+    if !asset.is_installed() {
+        return Err(AssetError::NotInstalled(asset.asset_id.clone()));
+    }
     let install_folder = asset
         .install_folder
         .clone()
-        .expect("existing install folder to be removed");
+        .expect("install_folder is specified, ensured in code before this.");
     let asset_path = installed_path(&install_folder);
     std::fs::remove_dir_all(asset_path)?;
